7-Minute Turn Based TicTacToe for Ether
Host a game, place a bet and get another to join your game. The winner gets the pot. If no one wins, the pot is split 50/50. If one player decides to back out have way through the game, the other player gets the pot and can claim their reward: there is roughly a 7 minute time period per turn, if that is broken, the other player is entitled to the pot.